requirements.txt
# Minimum Django and REST framework version
Django>=1.6
djangorestframework>=2.4.3
django-filter==0.15.3
jinja2>2.0
jsonschema>3.0
# Test requirements
pytest-django==3.8.0
pytest==5.4.1
pytest-cov==2.8.1
flake8==2.2.2

# wheel for PyPI installs
wheel==0.24.0

# MkDocs for documentation previews/deploys
mkdocs==0.11.1
